src/discord_backend/requirements.txt
discord.py
audioop-lts
httpx